create valutes final

+++ database/migrations/2022_11_30_215534_create_cbrf_table.php
@@ -16,7 +16,7 @@ class CreateCbrfTable extends Migration
         Schema::create('cbrf', function (Blueprint $table) {
             $table->id();
             $table->timestamps();
-            $table->float('usdrub');
+            $table->double('usdrub');
         });
     }
